backend/requirements.txt
# FastAPI and Server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
websockets==12.0

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0

# Database
motor==3.3.2
pymongo==4.6.1

# Computer Vision
opencv-python==4.9.0.80
mediapipe==0.10.9
numpy==1.26.3

# Utilities
pydantic==2.5.3
pydantic-settings==2.1.0
